+++ visualization/plot_profile.py
@@ -551,7 +551,7 @@ def smp_unlabelled(smp, smp_name, file_name="output/plots_data/smp_unlabelled.pn
     else:
         smp_wanted = smp_name
     smp_profile = smp[smp["smp_idx"] == smp_wanted]
-    ax = sns.lineplot(data=(smp_profile["distance"], smp_profile["mean_force"]))
+    ax = sns.lineplot(data=smp_profile, x="distance",y ="mean_force") #ax = sns.lineplot(data=(smp_profile["distance"], smp_profile["mean_force"]))
     plt.title("Unlabelled SMP Profile {}".format(smp_name))
     ax.set_xlabel("Snow Depth [mm]")
     ax.set_ylabel("Mean Force [N]")
@@ -582,7 +582,7 @@ def smp_labelled(smp, smp_name, title=None, file_name="output/plots_data/smp_lab
         smp_wanted = smp_name
 
     smp_profile = smp[smp["smp_idx"] == smp_wanted]
-    ax = sns.lineplot(data=(smp_profile["distance"], smp_profile["mean_force"]))
+    ax = sns.lineplot(data=smp_profile, x="distance",y ="mean_force") #ax = sns.lineplot(data=(smp_profile["distance"], smp_profile["mean_force"]))
     if title is None:
         plt.title("{} SMP Profile Distance (1mm layers) and Force\n".format(smp_name))
     else: